cleaned up google auth
