--- frontend/src/firebase/auth_google_provider_create.ts
@@ -1,5 +1,0 @@
-import { GoogleAuthProvider } from "firebase/auth";
-
-const provider = new GoogleAuthProvider();
-
-export {provider};
