+++ frontend/src/pages/Home.tsx
@@ -5,8 +5,6 @@ const Home: React.FC = () => {
     <div>
       <h1>Welcome to the Home Page</h1>
       <p>This is a temporary home page.</p>
-     
-
     </div>
   );
 };
